+++ autoblocks/human_review/__init__.py
@@ -1,16 +1,17 @@
-"""Human Review interface for reviewing and grading items."""
+"""Human Review API models."""
 
 from autoblocks._impl.human_review.models import BinaryScoreOptions
 from autoblocks._impl.human_review.models import Comment
 from autoblocks._impl.human_review.models import ContentType
 from autoblocks._impl.human_review.models import DiscreteRangeScoreOptions
-from autoblocks._impl.human_review.models import Field
 from autoblocks._impl.human_review.models import FieldComment
+from autoblocks._impl.human_review.models import FieldModel
 from autoblocks._impl.human_review.models import Grade
 from autoblocks._impl.human_review.models import Job
 from autoblocks._impl.human_review.models import JobItem
 from autoblocks._impl.human_review.models import JobItemDetail
 from autoblocks._impl.human_review.models import JobListItem
+from autoblocks._impl.human_review.models import JobsResponse
 from autoblocks._impl.human_review.models import Reviewer
 from autoblocks._impl.human_review.models import Score
 from autoblocks._impl.human_review.models import ScoreOptions
@@ -23,13 +24,14 @@ __all__ = [
     "Comment",
     "ContentType",
     "DiscreteRangeScoreOptions",
-    "Field",
+    "FieldModel",
     "FieldComment",
     "Grade",
     "Job",
     "JobItem",
     "JobItemDetail",
     "JobListItem",
+    "JobsResponse",
     "Reviewer",
     "Score",
     "ScoreOptions",
